+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/components/projectoverview/ProjectOverviewViewModel.groovy
@@ -1,6 +1,7 @@
 package life.qbic.portal.sampletracking.components.projectoverview
 
 import groovy.beans.Bindable
+import life.qbic.business.project.subscribe.Subscriber
 import life.qbic.datamodel.dtos.projectmanagement.Project
 import life.qbic.portal.sampletracking.communication.Topic
 import life.qbic.portal.sampletracking.resource.ResourceService
@@ -22,10 +23,12 @@ class ProjectOverviewViewModel {
 
     @Bindable ProjectSummary selectedProject
     @Bindable String generatedManifest
+    final Subscriber subscriber
 
-    ProjectOverviewViewModel(ResourceService<Project> projectResourceService, ResourceService<StatusCount> statusCountService){
+    ProjectOverviewViewModel(ResourceService<Project> projectResourceService, ResourceService<StatusCount> statusCountService, Subscriber subscriber){
         this.projectResourceService = projectResourceService
         this.statusCountService = statusCountService
+        this.subscriber = subscriber
         fetchProjectData()
         subscribeToResources()
     }
@@ -48,6 +51,7 @@ class ProjectOverviewViewModel {
         this.statusCountService.subscribe({updateSamplesReceived(it.projectCode, it.count)}, Topic.SAMPLE_RECEIVED_COUNT_UPDATE)
         this.statusCountService.subscribe({updateSamplesFailedQc(it.projectCode, it.count)}, Topic.SAMPLE_FAILED_QC_COUNT_UPDATE)
         this.statusCountService.subscribe({updateDataAvailable(it.projectCode, it.count)}, Topic.SAMPLE_DATA_AVAILABLE_COUNT_UPDATE)
+        this.statusCountService.subscribe({updateSamplesLibraryPrepFinished(it.projectCode, it.count)}, Topic.SAMPLE_LIBRARY_PREP_FINISHED)
     }
 
     private void addProject(Project project) {
@@ -71,6 +75,10 @@ class ProjectOverviewViewModel {
 
     private void updateSamplesFailedQc(String projectCode, int sampleCount) {
         getProjectSummary(projectCode).samplesQcFailed = sampleCount
+    }
+
+    private void updateSamplesLibraryPrepFinished(String projectCode, int sampleCount){
+        getProjectSummary(projectCode).samplesLibraryPrepFinished = sampleCount
     }
 
     /**