+++ sample-tracking-status-overview-domain/src/main/groovy/life/qbic/business/samples/count/CountSamples.groovy
@@ -78,6 +78,17 @@ class CountSamples implements CountSamplesInput{
     }
   }
 
+  @Override
+  void countLibraryPrepFinishedSamples(String projectCode) {
+    try {
+      sampleStatuses = dataSource.fetchSampleStatusesForProject(projectCode)
+      int libraryPrepFinished = countSamplesFromStatus(Status.LIBRARY_PREP_FINISHED)
+      output.countedLibraryPrepFinishedSamples(projectCode, sampleStatuses.size(), libraryPrepFinished)
+    } catch (Exception e) {
+      output.failedExecution(e.getMessage())
+    }
+  }
+
   private int countSamplesFromStatus(Status status) {
     int receivedIndex = statusesInOrder.indexOf(status)
     // statuses that are not considered in the ordered list return -1, meaning the sample is not counted